requirements.txt
numpy==1.19.2
pandas==1.1.3
Flask==1.1.2
flask_marshmallow==0.14.0
flask_sqlalchemy==2.4.4
Pillow==8.1.0
plotly==4.14.3
python_dateutil==2.8.1
qrcode==6.1
gunicorn